requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-engineio==4.10.1
python-socketio==5.11.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
prometheus-flask-exporter==0.23.0
redis==5.0.6
